refactor: reduce warning

+++ pvzclass/Classes/Zombie.cpp
@@ -553,7 +553,7 @@ void PVZ::Zombie::LandFlyer(PVZ::DamageFlags flag)
 void PVZ::Zombie::GetTrackPosition(const char* trackName, float& thePosX, float& thePosY)
 {
 	PVZ::Memory::WriteArrayUnsafe<const char>(PVZ::Memory::Variable + 100, trackName, strlen(trackName));
-	PVZ::Memory::WriteMemoryUnsafe<char>(PVZ::Memory::Variable + 100 + strlen(trackName), '\0');
+	PVZ::Memory::WriteMemoryUnsafe<char>(PVZ::Memory::Variable + 100 + DWORD(strlen(trackName)), '\0');
 	PVZ::Memory::Execute(AsmBuilder()
 		.mov_reg_imm(REG_EDI,PVZ::Memory::Variable)
 		.push_imm32(PVZ::Memory::Variable + 4)
